+++ packages/@plugin/plugin-babel/config/webpack.vue.cjs
@@ -1,21 +1,5 @@
 // const path = require("path");
 module.exports = {
-  rules: [
-    {
-      test: /\.(ts|js)$/,
-      include: [
-        {
-          __astType: "pathResolve",
-          args: ["./src"],
-        },
-      ],
-      exclude: [/node_modules/, /public/, /(.|_)min\.js$/],
-      use: [
-        {
-          loader: "babel-loader",
-        },
-      ],
-    },
-  ],
+  rules: [],
   plugins: [],
 };
