+++ packages/@plugin/plugin-babel/generator/index.cjs
@@ -47,54 +47,8 @@ const vueBabelConfig = {
   plugins: ["@vue/babel-plugin-jsx", ...commonBabelPlugins],
 };
 
-// 构建工具配置生成器映射
-const buildToolConfigGenerators = {
-  webpack: ({ test }) => {
-    const baseRule = {
-      test,
-      include: [
-        {
-          __astType: "pathResolve",
-          args: ["./src"],
-        },
-      ],
-      exclude: [/node_modules/, /public/, /(.|_)min\.js$/],
-      use: [{ loader: "babel-loader" }],
-    };
-
-    return {
-      rules: [baseRule],
-      plugins: [],
-    };
-  },
-  vite: ({ babelConfig }) => ({
-    plugins: [
-      {
-        name: "vite-plugin-babel",
-        transform: (code, id) => {
-          if (id.match(/\.(jsx?|tsx?)$/)) {
-            return require("@babel/core").transformSync(code, {
-              ...babelConfig,
-              filename: id,
-            }).code;
-          }
-        },
-      },
-    ],
-  }),
-  rollup: ({ babelConfig }) => ({
-    plugins: [
-      require("@rollup/plugin-babel")({
-        ...babelConfig,
-        extensions: [".js", ".jsx", ".ts", ".tsx"],
-      }),
-    ],
-  }),
-};
-
 module.exports = (generatorAPI, template, buildTool) => {
   let config;
-  let test;
   if (template === "react") {
     config = {
       babel: reactBabelConfig,
@@ -104,7 +58,6 @@ module.exports = (generatorAPI, template, buildTool) => {
         "@babel/preset-react": "^7.24.7",
       },
     };
-    test = /\.(ts|tsx|js|jsx)$/;
   } else if (template === "vue") {
     config = {
       babel: vueBabelConfig,
@@ -116,21 +69,7 @@ module.exports = (generatorAPI, template, buildTool) => {
         "@ant-design-vue/vue-jsx-hot-loader": "^0.1.4",
       },
     };
-    test = /\.(ts|js)$/;
   }
-
-  // 获取构建工具配置生成器
-  const configGenerator = buildToolConfigGenerators[buildTool];
-  if (!configGenerator) {
-    throw new Error(`不支持的构建工具: ${buildTool}`);
-  }
-
-  // 生成构建工具特定配置
-  const buildToolConfig = configGenerator({
-    test,
-    template,
-    babelConfig: config.babel,
-  });
 
   // 扩展package.json配置
   generatorAPI.extendPackage({
@@ -139,7 +78,8 @@ module.exports = (generatorAPI, template, buildTool) => {
 
   generatorAPI.protocolGenerate({
     [pluginToBuildToolProtocol.ADD_COMPILER_CONFIG]: {
-      config: buildToolConfig,
+      compiler: "babel",
+      template,
       buildTool,
     },
   });
