+++ packages/@plugin/plugin-swc/config/webpack.config.js
@@ -1,19 +1,3 @@
 module.exports = {
-  rules: [
-    {
-      test: /\.(ts|tsx|js|jsx)$/,
-      exclude: [/node_modules/, /public/, /(.|_)min\.js$/],
-      include: [
-        {
-          __astType: "pathResolve",
-          args: ["./src"],
-        },
-      ],
-      use: [
-        {
-          loader: "swc-loader",
-        },
-      ],
-    },
-  ],
+  rules: [],
 };
